refactor: move key_to_bytes from ssh/pty.rs to term_input.rs
