+++ crates/omnyssh/src/lib.rs
@@ -7,6 +7,7 @@ pub mod app;
 pub mod config;
 pub mod event;
 pub mod ssh;
+pub mod term_input;
 pub mod ui;
 pub mod update;
 pub mod utils;
